Initial Import
CourseFit is a mock-up, using PHP, Javascript and CSS, of a University
of Toronto online course recommendation engine, based on Twitter
Bootstrap

It gives a general idea of what the interface will look like - achieving
cleanness, simplicity and user-friendliness in one. This is completed as
a part of the Senior Year Capstone Design project.